split gateways
