+++ .github/workflows/python-package.yml
@@ -26,7 +26,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         find . -name 'requirements*.txt' -exec python -m pip install -r {} \;
-    # add flake8 above before enabling
     # - name: Lint with flake8
     #   run: |
     #     # stop the build if there are Python syntax errors or undefined names
